+++ lib/dargres.dart
@@ -1,5 +1,6 @@
 library dargres;
 
+export 'src/execution_context.dart';
 export 'src/core.dart';
 export 'src/connection_settings.dart';
 export 'src/query.dart';
@@ -9,5 +10,7 @@ export 'src/ssl_context.dart';
 export 'src/transaction_context.dart';
 export 'src/exceptions.dart';
 export 'src/to_statement.dart';
+export 'src/pool/postgresql_pool.dart';
+