+++ test/terrier_buffer_reader_test.dart
@@ -6,7 +6,7 @@ import 'package:dargres/src/dependencies/buffer_terrier/raw_reader.dart';
 import 'package:test/test.dart';
 
 import 'package:dargres/src/dependencies/buffer_terrier/debug_hex_codec.dart';
-import 'package:matcher/matcher.dart';
+//import 'package:matcher/matcher.dart';
 
 import 'package:dargres/src/dependencies/buffer_terrier/raw_data.dart';
 import 'package:dargres/src/dependencies/buffer_terrier/raw_value.dart';