+++ test/terrier_buffer_writer_test.dart
@@ -3,7 +3,7 @@ import 'package:dargres/src/dependencies/buffer_terrier/raw_writer.dart';
 import 'package:test/test.dart';
 
 import 'package:dargres/src/dependencies/buffer_terrier/debug_hex_codec.dart';
-import 'package:matcher/matcher.dart';
+//import 'package:matcher/matcher.dart';
 
 import 'package:dargres/src/dependencies/buffer_terrier/raw_data.dart';
 import 'package:dargres/src/dependencies/buffer_terrier/raw_value.dart';